Create bmi.py
Create a class Person with the below attributes:
name - String
height - Number
weight - Number
bmi - Number
bmi_status - String
Create the __init__ method which takes the parameters name, height and weight in the above sequence. The method should set the value of attributes to parameter values. Also __init__ method should set default values of bmi and bmi_status attribute as 0 respectively.
Create a method inside the Person class with the name calculateBmi. This method will set the value of bmi of the calling object. This method takes no argument and it calculates the bmi, it also sets the value of bmi of Person object.
No return value is expected from this method, it will just set the bmi of the Person object.
Formula to be used for bmi calculation is: bmi = weight/(height*height)
Round the calculated values to nearest integer.

Create another class Society with the below two attributes:
1. bmi_status of type dictionary having the bmi status (a string value) as the key and a tuple having two numbers (lower bmi limit and upper bmi limit representing lower and upper bmi limits for the particular bmi status) as value.
Note: The value for the upper bmi limit cannot be lower than the lower bmi limit. 
Code should have a check on input data and accordingly set the values in case upper and lower limits are given in reverse order. 
2.person_list of type List having Person objects. 
Create the __init__ method which takes all parameters in the above sequence. The method should set the value of attributes to parameter values inside the method.
Create 2 methods inside the Society class:
calculateBmiAndStatusByName.
The method takes a string representing the name of a person as an argument. 
The method should find the respective person from the person_list of the Society whose name is passed as the argument and calculate the bmi value for that person. Method will also find the bmi status based on the bmi value calculated. Value for the bmi status is available in the bmi_status dictionary of the Society. Finally, the method will update the bmi and the bmi_status of the person and return true. The search of Person by name should be case insensitive.
If the person with given name is not found in the person_list, then the method should return false.
Note: Use the method calculateBmi defined in the Person class to calculate the bmi of the person and set the value for the attribute bmi.
All comparisons are case insensitive.
If more than one person found with the same name given as argument, method will update the bmi and bmi_status for all those persons.
removelnvalidPersons:
The method takes a number representing the value of bmi of a person as an argument. The method should find the persons from the person_list of the Society, whose bmi is less than the value passed as the argument and return those Persons as a list.

Instructions to write main section of the code
Create the respective objects (Person and Society) with the given sequence of arguments to fulfill the init method requirement define in the respective classes referring to the below instructions.
1)	 Create a list of Person objects. To create the list,
a.	Read a number representing how many person objects to be created.
b.	Read data related to attributes of Person (i.e. name, height and weight). Create the Person object and add the object to the list of Person objects which will be provided to the Society object. This point repeats for the number of Person objects to be created 
2)	 Create a dictionary (key:value pairs) having the bmi_status(a string) as key and range of bmi(a tuple having two numbers representing the lower limit and the upper limit of bmi) as value. To create the dictionary:
a.	Read a number for how many elements to be there in the dictionary.
b.	Read a String representing the bmi_status(key) and two numbers (the lower limit and the upper limit) to create a tuple as the value and add the key:value pair in the dictionary. This point repeats for the number of elements (key:value pairs) to be added to the dictionary 
3)	 Create Society object by passing the List of Person objects (created and as mentioned in above point) and the dictionary created in point above as the arguments. 
4)	Call the method calculateBmiAndStatusByName mentioned above from the main section.
5)	If the method calculateBmiAndStatusByName returns True. Display the bmi and the bmi_status of the person (from the person list of the Society object created in point above) for those Person's whose bmi value is not 0 and whose name is supplied as argument to the function. 
6)	If the method calculateBmiAndStatusByName returns False, then display the message "No Person Exists".
7)	Read a number of bmi value to be taken as an argument for method removeInvalidPersons
8)	Call the method removelnvalidPersons method above from the main section
9)	Display the name and the bmi of all of the persons in the list returned by the removelnvalidPersons of the Society object created in point above.