Create bloodgroup.py
Create class Blood with attributes:
bloodGroup of type String
unitInHand of type Number
Create the constructor which takes parameters in above sequence for bloodGroup and unitInHand.
Method will set the values passed as argument to the attributes of the Blood object created.
Note: Blood group should be in upper case (eg. A,B+,O etc.)
Create a class BloodBank with attributes:
bloodList is a list of Blood objects
create the constructor which takes as parameter a list of blood objects. Method will set the value passed as argument to the attribute of the BloodBank object created.
Define two methods as below in the BloodBank class:
1)	isBloodAvailable : This method will take as argument a string representing the blood group and a number representing the units of blood required. Method will check whether the required unit of blood of the input blood group is available or not in the bloodList of the BloodBank. Method will return True if required unit of blood of given blood group is available in the blood bank. If blood of the blood group passed as an argument is not available or Blood of the required blood group is not available in the required quantity, method will return False.
2)	findMinBloodStock : This method will find the bloodGroup of the Blood in the bloodList having minimum unit in hand. In case there are more than one Blood objects having the minimum value for unitInHand, method will return all the bloodGroups found.

These methods should be called from the main section.

Instruction to define the main function:
Create the respective objects (Blood and BloodBank) with the given sequence of arguments to fulfill the __init__ method requirement defined in the respective classes referring to the below instructions:
1)	Create a list of blood objects which will be provided to the BloodBank object to be created. To create the list-
A)	Read a number for the count of Blood objects to be created and added to the list;
B)	Create a Blood object after reading the data related to it and add the objects to the list of Blood objects. This point repeats for the number of Blood objects to be created (considered in the first line of input)
2)	Create BloodBank object by passing the list of Blood objects as the arguments. 
3)	Read a string value representing the blood group to be provided as argument to the method isBloodAvailable
4)	Read a number value representing the unit required to be passed as argument to the method isBloodAvailable.
5)	Call the method isBloodAvailable by passing the values read in the points above from main section.
6)	Display the message “Blood Available.” (excluding the quotes) if method isBloodAvailable returns True. If the method isBloodAvailable returns False, print the message “Blood not available.” (excluding the quotes)
7)	Call the method findMinBloodStock and print the blood groups returned in separate lines.