Create movie.py
Create a class Movie with the below attribute:
movie_id of type number
movie_name of type string
ticket_cost of type number 
cost_category represents the ticket cost category of type string and this attribute should have default value of “Default”
Create the __init__ method which takes movie_id, movie_name, ticket_cost as parameters in the given sequence.
Create a method inside the Movie class with the name assign Price_category. This method assigns the value to the category attribute of the Movie based on the below category:
If the cost of the ticket is greater than 0 and less than 150, then the category will be General.
If the cost of the ticket is greater than and equal to 150 and less than 250 then the category will be silver.
If the cost of the ticket is greater than or equal to 250 and less than 350 then category will be gold.
If the cost of the ticket is greater than and equal to 350, then the category will be platinum.
Create another class Movie_ticket with the below attributes:
 customer_name of type string represents the name of the Person who has requested for booking
movie_name of type string represents the name of the movie the ticket is booked for.
viewerCount of type Number represents the count of viewers.
totalTicketCost of type number represents the total cost of ticket.
Create the __init__ method which takes all parameter in the above sequence
Create the method as below:-
getCategoryWiseCount  which is not a method of any class. The method should take as argument a list of Movie objects and return a dictionary consisting of category: countOfMovies as key:value pair where category represents the category of a Movie and countOfMovies represents the count of movies of that category calculated from the Movie list passed as argument after assigning them to their respective category.
bookMovieTicket which is not a method of any class. This method will take as argument a list of movie objects and customer name, movie name and count of viewers. Method will return the Movie_Ticket created will find the movie object from the list passed as first argument having the name equal to the movie name passed as the third argument and find the ticket_cost of the Movie.
Method will calculate the totalTicketCost by using the formula:
totalticketCost= ticket_cost*viewer_count
Method will finally and create and return the Movie_Ticket with the values passed as argument for customer name, movie name, viewer count and total ticket cost calculated. If the movie with the movie name passed as argument is not found in the Movie list passed as argument, method returns None.



Instruction to write Main section: -
Create the respective objects (Movie, Movie_Ticket) with the given sequence of arguments to fulfill the__init__ method requirement defined in the respective classes referring to the below instructions-
1)	Input the number of movie objects that needs to be created.
2)	Create the Movie object after reading the data related to it and add the objects to the list of movie objects which will be provided to the methods getCategoryWiseCount and bookMovieTicket. This point repeats for the number of Movie objects to be created that is given in the first line of input.
3)	Read a string to represent the customer name to be passed as argument to the method bookMovieTicket.
4)	Read a string to represent the movie name to be passed as argument to method bookMovieTicket
5)	Read a number representing the count of viewers to be passed as argument to the method bookMovieTicket
6)	Call the method getCategoryWiseCount   defined above from main section by passing the list created in above point.
7)	Display a message “getCategoryWiseCount  “
8)	Display the result value returned by the method getCategoryWiseCount   in the method description.
9)	Call the method bookMovieTicket defined above from main section by passing the list created in above point and value read in above points.
10)	Display the customer name and totalticketCost of the Movie_Ticket object returned by the method bookMovieTicket
11)	If the method bookMovieTicket returns None, print the message “Movie not available.”